+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### vNext
 
-- 
+- Add support of wallet export as a JSON backup file
 
 ### v1.36.0